+++ CHANGELOG.md
@@ -13,8 +13,12 @@ All notable changes to this project will be documented in this file.
 * Support for Linux.
 * Tests that run on GitHub Actions.
 * Package typings when installing with `pip`.
+* Centralized logging, `--verbose` and `--silent` flags.
+* Ability to turn off color output with `--no-color` flag or `NO_COLOR=1` environment variable.
 
 ### Changed
+
+* Removed space from title text.
 
 ### Fixed